% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geoms.R
\name{add_geom_usage_to_files}
\alias{add_geom_usage_to_files}
\title{add_geom_usage_to_files}
\usage{
add_geom_usage_to_files(data_file_info)
}
\description{
\code{add_geom_usage_to_files()} is the main function you'll use to add geom usage
data to your files.
}
