% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_geom_usage}
\alias{data_example_geom_usage}
\title{Core geoms dataset}
\format{
\subsection{\code{data_geoms}}{

A data frame with 57 rows and 4 columns:
\describe{
\item{geom_name}{Name of geom}
\item{package_name}{Name of package containing geom}
\item{docs_title}{Title of geom documentation page}
\item{docs_description}{Description of geom documentation page}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
data_example_geom_usage
}
\description{
A dataset containing example geom usage
}
\keyword{datasets}
