% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geoms.R
\name{extract_geom_arguments}
\alias{extract_geom_arguments}
\title{extract_geom_arguments}
\usage{
extract_geom_arguments(geom_call)
}
\arguments{
\item{geom_call}{A single call of a geom, provided as a string.}
}
\description{
\code{extract_geom_arguments()} extracts and labels arguments as aes() or not from
a single geom call
}
