% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fill_geom_usage_date_data}
\alias{fill_geom_usage_date_data}
\title{Fill geom usage per day}
\usage{
fill_geom_usage_date_data(data_usage_per_day, target_year = 2025)
}
\arguments{
\item{data_usage_per_day}{Tibble containing geom usage per day, generated from
\code{summarise_per_day()}}

\item{target_year}{Target year. Defaults to 2025 as that's the first year of
ggplot2wraped!}
}
\description{
\code{fill_geom_usage_date_data()} utility function for filling geom usage per day
from the function \code{summarise_per_day()}
}
