% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fill_geom_usage_date_data_with_nesting}
\alias{fill_geom_usage_date_data_with_nesting}
\title{Fill geom usage per day with nesting}
\usage{
fill_geom_usage_date_data_with_nesting(
  data_usage_per_day,
  nesting_column = NULL,
  target_year = 2025
)
}
\arguments{
\item{data_usage_per_day}{Tibble containing geom usage per day, generated from
\code{summarise_per_day()}}

\item{target_year}{Target year. Defaults to 2025 as that's the first year of
ggplot2wraped!}
}
\description{
\code{fill_geom_usage_date_data_with_nesting()} utility function for filling geom usage per day
from the function \code{summarise_per_day()} supporting nesting.
}
