% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_info.R
\name{get_code_file_info}
\alias{get_code_file_info}
\title{Get file info}
\usage{
get_code_file_info(paths, file_type = c(".R", ".qmd", "Rmd"))
}
\arguments{
\item{paths}{A singular path or a vector of file paths to look in for code
files}

\item{file_types}{Which R code files to look at. Defaults to R, Quarto and
RMarkdown.}
}
\value{
A tibble.
}
\description{
\code{get_code_file_info()} provides a formatted tibble containg file info about
your selected code files.
}
