% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_info.R
\name{get_docs_details}
\alias{get_docs_details}
\title{Get docs details}
\usage{
get_docs_details(geom_name, package_name, section)
}
\arguments{
\item{package_name}{Package name}

\item{section}{Section heading of docs page}

\item{get_docs_details}{Name of geom}
}
\value{
A tibble.
}
\description{
\code{get_docs_details()} extracts sections of a function documentation page as
text
}
