% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geoms.R
\name{get_geoms_from_code_file}
\alias{get_geoms_from_code_file}
\title{get_geoms_from_code_file}
\usage{
get_geoms_from_code_file(file_path, geoms_dataset)
}
\arguments{
\item{file_path}{Path to a code file.}

\item{geoms_dataset}{A data.frame containing a column 'geom_name' with the
name of a geom of interest.}
}
\value{
A tibble.
}
\description{
\code{get_geoms_from_code_file()} uses \code{{astgrepr}} to both extract and count geom
usage from a code file.
}
