% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_info.R
\name{get_geoms_from_package}
\alias{get_geoms_from_package}
\title{Get geoms from a package}
\usage{
get_geoms_from_package(package_name)
}
\arguments{
\item{package_name}{Package name, as a string}
}
\value{
A tibble.
}
\description{
\code{get_geoms_from_package()} returns a data.frame containing all geoms from a
package.
}
