% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geoms.R
\name{get_tidytuesday_repo_geom_usage}
\alias{get_tidytuesday_repo_geom_usage}
\title{get_tidytuesday_repo_geom_usage}
\usage{
get_tidytuesday_repo_geom_usage(
  repo_owner,
  repo_name,
  branch_name = "main",
  data_geoms = ggplot2wrapped::data_geoms,
  report_type = "simple"
)
}
\description{
\code{get_tidytuesday_repo_geom_usage()} get geom usage from a github repo (ideally
a tidytuesday rpeo)
}
