% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2wrapped-package.R
\docType{package}
\name{ggplot2wrapped-package}
\alias{ggplot2wrapped}
\alias{ggplot2wrapped-package}
\title{ggplot2wrapped: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
