% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ggplot-wrapped.R
\name{ggplot_wrapped_2025}
\alias{ggplot_wrapped_2025}
\title{Get file info}
\usage{
ggplot_wrapped_2025(
  paths,
  file_type = c(".R", ".qmd", "Rmd"),
  report_type = c("scrollytelling"),
  year_wrapped = 2025,
  file_action = "created",
  exported_report = "your-ggplot2-wrapped.html"
)
}
\arguments{
\item{paths}{A singular path or a vector of file paths to look in for code
files}

\item{file_types}{Which R code files to look at. Defaults to R, Quarto and
RMarkdown.}
}
\value{
A tibble.
}
\description{
\code{get_code_file_info()} provides a formatted tibble containg file info about
your selected code files.
}
