% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{make_aes_type_percentile_highcharts}
\alias{make_aes_type_percentile_highcharts}
\title{Interactive aes usaege chart}
\usage{
make_aes_type_percentile_highcharts(data_geom_usage)
}
\arguments{
\item{data_geom_usage}{Tibble containing geom usage data produced via
\code{add_geom_usage_to_files()}.}
}
\value{
A ggplot2 chart
}
\description{
\code{make_aes_type_percentile_highcharts()} interactive bar chart showing geoms
that use both dataviz-level and chart-level aesthetics and their relative \%
}
