% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-charts.R
\name{make_geom_icon_usage_grid}
\alias{make_geom_icon_usage_grid}
\title{Grid of geom groupings and their usage}
\usage{
make_geom_icon_usage_grid(data_geom_usage, show_which_geoms = "used")
}
\arguments{
\item{data_geom_usage}{Tibble containing geom usage data produced via
\code{add_geom_usage_to_files()}.}

\item{show_which_geoms}{Which geoms are shown? Default is "used" which
highlights only used geoms. Providing "all" will show all geoms.
RMarkdown.}
}
\value{
A ggplot2 chart
}
\description{
\code{make_geom_icon_usage_grid()} creates a ggplot2 with a grid of geom groupings
from the official {ggplot2} charts, optionally highlighting which geoms are
used.
}
