% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-charts.R
\name{make_geom_upset_chart}
\alias{make_geom_upset_chart}
\title{Get file info}
\usage{
make_geom_upset_chart(data_geom_usage, focus = c("geom", "date"))
}
\description{
\code{summarise_per_day()} utility function
}
