% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarise_ggplot2_geom_usage}
\alias{summarise_ggplot2_geom_usage}
\title{Overall summary of ggplot2 geom usage}
\usage{
summarise_ggplot2_geom_usage(data_geom_usage)
}
\arguments{
\item{data_geom_usage}{Tibble containing geom usage data produced via
\code{add_geom_usage_to_files()}.}
}
\description{
\code{summarise_ggplot2_geom_usage()} summarises use of the standard geoms in the
{ggplot2} package
}
