% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarise_per_day}
\alias{summarise_per_day}
\title{Summarise geom usage per day}
\usage{
summarise_per_day(data_geom_usage, measure = "per_day_individual_geom_usage")
}
\arguments{
\item{data_geom_usage}{Tibble containing geom usage data produced via
\code{add_geom_usage_to_files()}.}

\item{measure}{How to summatrise usage by day? Defaaults to "per_day_individual_geom_usage",
can also be one of; "per_day_files_with_geoms", "per_day_unique_geoms",
"per_day_total_geom_usage"}
}
\description{
\code{summarise_per_day()} utility function for summarising geom usage by several
different metrics
}
