% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarise_per_file}
\alias{summarise_per_file}
\title{Summarise geom usage per file}
\usage{
summarise_per_file(data_geom_usage)
}
\arguments{
\item{data_geom_usage}{Tibble containing geom usage data produced via
\code{add_geom_usage_to_files()}.}
}
\description{
\code{summarise_per_file()} utility function for summarising geom usage by several
different metrics
}
