% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarise_top_4_ggplot_geom_usage}
\alias{summarise_top_4_ggplot_geom_usage}
\title{Overall summary of ggplot2 geom usage}
\usage{
summarise_top_4_ggplot_geom_usage(data_geom_summary, summarise_by = "geoms")
}
\arguments{
\item{data_geom_summary}{Tibble containing ggplot2's geom usage data produced via
\code{summarise_ggplot2_geom_usage()}.}
}
\description{
\code{summarise_ggplot2_geom_usage()} summarises use of the standard geoms in the
{ggplot2} package
}
