% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-charts.R
\name{fill_geom_usage_date_data}
\alias{fill_geom_usage_date_data}
\title{Get file info}
\usage{
fill_geom_usage_date_data(data_usage_per_day, target_year = 2025)
}
\description{
\code{fill_geom_usage_date_data()} utility function
}
