% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-charts.R
\name{fill_geom_usage_date_data_with_nesting}
\alias{fill_geom_usage_date_data_with_nesting}
\title{Get file info}
\usage{
fill_geom_usage_date_data_with_nesting(
  data_usage_per_day,
  nesting_column = NULL,
  target_year = 2025
)
}
\description{
\code{fill_geom_usage_date_data_with_nesting()} share doc page
}
