% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{make_aes_type_percentile_highcharts}
\alias{make_aes_type_percentile_highcharts}
\title{Get file info}
\usage{
make_aes_type_percentile_highcharts(data_geom_usage)
}
\arguments{
\item{paths}{A singular path or a vector of file paths to look in for code
files}

\item{file_types}{Which R code files to look at. Defaults to R, Quarto and
RMarkdown.}
}
\value{
A tibble.
}
\description{
\code{get_code_file_info()} provides a formatted tibble containg file info about
your selected code files.
}
