% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-charts.R
\name{make_geom_icon_usage_grid}
\alias{make_geom_icon_usage_grid}
\title{Get file info}
\usage{
make_geom_icon_usage_grid(data_geom_usage, show_which_geoms = c("used", "all"))
}
\arguments{
\item{paths}{A singular path or a vector of file paths to look in for code
files}

\item{file_types}{Which R code files to look at. Defaults to R, Quarto and
RMarkdown.}
}
\value{
A tibble.
}
\description{
\code{get_code_file_info()} provides a formatted tibble containg file info about
your selected code files.
}
