% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{make_geom_interactive_upset_chart}
\alias{make_geom_interactive_upset_chart}
\title{Get file info}
\usage{
make_geom_interactive_upset_chart(
  data_geom_usage,
  height = NULL,
  min_geom_appearances = 10,
  min_interaction_size = 1
)
}
\description{
\code{summarise_per_day()} utility function
}
