% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{make_geom_interactive_venn_diagram}
\alias{make_geom_interactive_venn_diagram}
\title{Get file info}
\usage{
make_geom_interactive_venn_diagram(data_geom_usage)
}
\description{
\code{summarise_per_day()} utility function
}
