% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{make_geom_usage_pies_highcharts}
\alias{make_geom_usage_pies_highcharts}
\title{Interactive pie chart of geom usage}
\usage{
make_geom_usage_pies_highcharts(
  data_geom_usage,
  pie_measure_type = "geom_total_usage"
)
}
\arguments{
\item{data_geom_usage}{Tibble containing geom usage data produced via
\code{add_geom_usage_to_files()}.}

\item{pie_measure_type}{How should pie chart summarise geom used? Default to
"geom_total_usage", can also be "geom_usage_in_files"}
}
\value{
A ggplot2 chart.
}
\description{
\code{make_geom_usage_pies_highcharts()} produces an interactive pie chart showing
your top 4 geom usage across all calls or files
}
