% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-charts.R
\name{summarise_per_day}
\alias{summarise_per_day}
\title{Get file info}
\usage{
summarise_per_day(
  data_geom_usage,
  measure = c("per_day_individual_geom_usage", "per_day_files_with_geoms",
    "per_day_unique_geoms", "per_day_total_geom_usage")
)
}
\description{
\code{summarise_per_day()} utility function
}
